CR feedback: Move BlazorApi.cs to shared; use it from Analyzers test
