+++ src/Microsoft.AspNetCore.Blazor.Analyzers/ComponentParametersShouldNotBePublicAnalyzer.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using Microsoft.AspNetCore.Blazor.Shared;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -43,7 +44,7 @@ namespace Microsoft.AspNetCore.Blazor.Analyzers
 
             var parameterAttribute = declaration.AttributeLists
                 .SelectMany(list => list.Attributes)
-                .Where(attr => semanticModel.GetTypeInfo(attr).Type?.ToDisplayString() == "Microsoft.AspNetCore.Blazor.ParameterAttribute")
+                .Where(attr => semanticModel.GetTypeInfo(attr).Type?.ToDisplayString() == BlazorApi.ParameterAttribute.FullTypeName)
                 .FirstOrDefault();
 
             if (parameterAttribute != null && IsPublic(declaration))
