+++ src/Microsoft.AspNetCore.Blazor.Razor.Extensions/Microsoft.AspNetCore.Blazor.Razor.Extensions.csproj
@@ -38,4 +38,8 @@
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
   </ItemGroup>
+
+  <ItemGroup>
+    <Compile Include="..\shared\BlazorApi.cs" Link="shared\%(Filename)%(Extension)" />
+  </ItemGroup>
 </Project>
